fix: :bug: Fixed unique key bug for chapters and removed some commented out code
