+++ manga-finder/src/App.tsx
@@ -7,7 +7,6 @@ import "./index.css";
 import Browse from "./pages/Browse";
 import Error from "./pages/Error";
 import Home from "./pages/Home";
-import Library from "./pages/Library";
 import Manga from "./pages/Manga";
 import Reader from "./pages/Reader";
 
@@ -17,13 +16,11 @@ export default function App() {
   return (
     <div className="w-full">
       <NavBar />
-      {/* <FilterBar /> */}
       <div className="mt-12 lg:mt-16 mb-12 lg:mb-0">
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/browse" element={<Browse />} />
           <Route path="/browse/:query" element={<Browse />} />
-          <Route path="/library" element={<Library />} />
           <Route path="/manga/:mangaid" element={<Manga />} />
           <Route path="/reader/:chapterid" element={<Reader />} />
           <Route path="*" element={<Error />} errorElement={<Error />} />
@@ -33,6 +30,3 @@ export default function App() {
     </div>
   );
 }
-//Store
-//Use browser router in main.tsx
-//Failsafe if data doesn't load (404 for manga)
