+++ manga-finder/src/services/ChapterQueries.ts
@@ -1,12 +1,9 @@
-//IChapterResponse request string below
-// https://api.mangadex.org/at-home/server/{chapter.id}
-
 import { useQuery } from "@tanstack/react-query";
 import axios from "axios";
 
 export function getChapterData(mangaID: string) {
     const { isLoading, error, data } = useQuery({
-        queryKey: [mangaID],
+        queryKey: [mangaID + "ChapterQuery"],
         queryFn: () =>
             axios
                 .get(
@@ -16,5 +13,3 @@ export function getChapterData(mangaID: string) {
     });
     return { isLoading, error, data };
 }
-
-/* https://uploads.mangadex.org/data/3303dd03ac8d27452cce3f2a882e94b2/1-f7a76de10d346de7ba01786762ebbedc666b412ad0d4b73baa330a2a392dbcdd.png */
